Strip jolt-core into jolt-prover-legacy (#1632)
* first few strip steps

* stripped down to a jolt-prover

* clean up migration

* chore: finish jolt-prover migration

* fix: address CI migration failures

* fix: satisfy taplo formatting

* fix: make wasm verifier arithmetic word-size independent

* docs: clarify claim source taxonomy

* refactor: rename stable prover crate to legacy

* fix: replace unsafe Dory commitment transmute with safe conversion

`prover_dory_commitment_into_verifier` used `transmute_copy` between
`ArkGT` and `Bn254GT`, which silently relies on identical memory layout
and reads past the source if the wrappers ever diverge. Both wrap the
same patched `ark_bn254::Fq12`, so convert through the shared inner
element via a new symmetric `From<Fq12> for Bn254GT`. Zero runtime cost,
no `unsafe`, and divergence now becomes a compile error rather than
silent corruption on the soundness-critical commitment path.

Also stop the SDK verifier fixture test from passing vacuously: it now
fails loudly when fixtures are missing under CI, and gen-fixtures.sh
errors out instead of warn-and-continue when `fibonacci --save` does not
produce the expected files.


---------

+++ crates/jolt-claims/src/protocols/jolt/formulas/bytecode.rs
@@ -632,12 +632,14 @@ where
     );
 
     for (stage_value, stage_cycle_point) in stage_values.iter_mut().zip(inputs.stage_cycle_points) {
-        *stage_value *= eq_index_msb(stage_cycle_point, cycle_index);
+        *stage_value *= eq_index_msb(stage_cycle_point, cycle_index as u128);
     }
 
     let identity = IdentityPolynomial::new(inputs.r_address.len()).evaluate(inputs.r_address);
-    let spartan_outer_raf = identity * eq_index_msb(inputs.stage_cycle_points[0], cycle_index);
-    let spartan_shift_raf = identity * eq_index_msb(inputs.stage_cycle_points[2], cycle_index);
+    let spartan_outer_raf =
+        identity * eq_index_msb(inputs.stage_cycle_points[0], cycle_index as u128);
+    let spartan_shift_raf =
+        identity * eq_index_msb(inputs.stage_cycle_points[2], cycle_index as u128);
     let entry = if address_index == inputs.entry_bytecode_index && cycle_index == 0 {
         F::one()
     } else {
